requirements.txt
# Telegram Bot Framework
aiogram==3.4.1

# Database
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
SQLAlchemy==2.0.25

# Fuzzy String Matching
thefuzz==0.22.1
python-Levenshtein==0.23.0

# Logging
logging
pip install schedule
pip install psycopg2-binary python-dotenv schedule